pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "skill-router"
version = "0.1.0"
description = "Skill Router - Hackathon project"
requires-python = ">=3.11"
dependencies = [
    "pyyaml",
    "networkx",
    "anthropic>=0.40",
    "claude-agent-sdk>=0.1",
    "google-genai>=1.0",
    "mcp",
    "fastapi",
    "uvicorn[standard]",
    "python-multipart",
    "matplotlib",
    "pandas",
    "neo4j>=5.20",
]

[project.optional-dependencies]
bench = [
    "sentence-transformers>=3.0",
    "numpy>=1.26",
]
dev = [
    "pytest",
    "pytest-asyncio",
    "testcontainers[neo4j]>=4.0",
    "sentence-transformers>=3.0",
    "numpy>=1.26",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = ["integration: integration tests requiring docker/neo4j"]

[tool.hatch.build.targets.wheel]
packages = ["src/skill_router", "bench"]
